refactor!: move credential id helpers to common
